# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/quote": {
            "get": {
                "tags": [
                    "Quote"
                ],
                "summary": "Returns random quote (dev version)",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/story": {
            "get": {
                "tags": [
                    "Story"
                ],
                "summary": "Returns story for today",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/story/rate": {
            "post": {
                "tags": [
                    "Story"
                ],
                "summary": "Rate story",
                "parameters": [
                    {
                        "description": "Story rate",
                        "name": "rate",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "description": "Story ID",
                        "name": "ID",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/story/read": {
            "post": {
                "tags": [
                    "Story"
                ],
                "summary": "Read story",
                "parameters": [
                    {
                        "description": "Story ID",
                        "name": "ID",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        }
    }
}